ci: .buildkite: use encrypted `DOCUMENTER_KEY` and sign pipeline
* The `DOCUMENTER_KEY` environment variable is an encrypted version of
`DOCUMENTER_KEY` (i.e., an SSH-key generated by running
`DocumenterTools.genkeys(user="SciML", repo="SciMLDocs"))`),
encrypted using `bin/encrypt_variable` from
`cryptic-buildkite-plugin`.
Since it is an encrypted value, it's alright to have it present in the
actual workflows in the repository - this allows us to better enforce
step-specific `env` variable exposure.

* Refactor the `pipeline.yml` workflow into `aggregate.yml`. This will
allow us to can make `aggregate.yml`, which deals with sensitive env
variables, a signed pipeline, and appropriately annotate that in the
`pipeline.yml` that gets first uploaded when a job is started.

* Sign the `aggregate.yml` pipeline with `bin/sign_treehashes` from
`cryptic-buildkite-plugin`.

* `GITHUB_TOKEN` is used for authentication with GHA. Since we're
running this workflow on buildkite instead, this is no longer relevant.
So, remove it.