.buildkite: optimize S3 uploads with `--delete` and `--size-only` flags
Optimize S3 uploads of the aggregated multi-documentation that we do
with the `--delete` and `--size-only` flags.

* `--size-only`: Makes the size of each key the only criteria used to
decide whether to sync from source to destination. Without this, since
our timestamps are always newer (because we just copied all the files
over) we're copying everything up every time. Since the sizes of things
should be changing when uploading, this is fairly safe to set.

* `--delete`: Files that exist in the destination but not in the source
are deleted during sync. We want to remove files that no longer exist
locally, so that when we delete a file, it gets deleted on S3 as well.