.buildkite: aggregate: consolidate docs build and upload into one step
Buildkite enforces an in-built limit of 5GB on the size of artifacts
that can be uploaded by a build step, which we easily surpass [1].
In order to modify this limit, we would have to configure our own
artifact upload destination (S3 or others). But at that point, we would
be redundantly storing things in our S3 buckets (or whichever other
destination) and wasting build time zipping and unzipping these builds.
Instead, consolidate the build and upload step ,i.e., directly build the
documentation in the sandbox and upload it to S3, all in one step.

[1] https://buildkite.com/julialang/scimldocs/builds/79#0183da18-fbde-40c1-94aa-ea5f20f03d64/493-509