ci(GHA): properly set up CompatHelper workflow
The CompatHelper workflow, originally added in 0dc3ff2, was not added in
accordance with the convention the GitHub Actions expects[^1], i.e.,
instead of being in a YAML file in the `workflows` directory, it was in a
file named `workflows` itself. This caused the CompatHelper workflow to
never run.
Fix this by moving the workflow into `.github/workflows/CompatHelper.yml`.

Fixes: 0dc3ff2 ("doc compat")

[^1]: https://docs.github.com/en/actions/quickstart#creating-your-first-workflow