ci: perform only shallow checkouts of the repository
Given the size of the repository, we spend an inordinate time in CI for
each step just cloning the repository (sometimes upto ~7 mins[^1]);
an inefficiency that is obviously problematic.
Fix this, by performing only a shallow checkout of the repository using
the `hasura/smooth-checkout` plugin. The steps defined in the WebUI have
also been updated to do the same. This brings down the repository
checkout time significantly, to just 1-2s[^2].

Since the documentation workflow will require switching branches and
pushing to the repository, additionally fetch just the `HEAD` of
`gh-pages` in that workflow before actually building and trying to deploy
the documentation.

[^1]: https://buildkite.com/julialang/scimldocs/builds/2955#018bdc82-534f-4aa9-8e3e-00865db3fa22/56
[^2]: https://buildkite.com/julialang/scimldocs/builds/2961#018bdcc2-45f7-45a2-b600-8cff02d2360b/101